the script that checks status of the ad
This function checks whether the ads are active or not and label them accordingly. With crawling function, we can have (if executed daily) a folder full of CSV files with each of them corresponding to each day. If they are added together we have our updated data pool and this function should be applied to that data pool regularly which increase in size, again daily.